requirements.txt
django==3.1.3
djangorestframework==3.11.0
django-cors-headers==3.4.0
markdown==3.3.3
django-filter==2.4.0
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0